requirements.txt
discord.py 
feedparser 
requests 
google-api-python-client